feat : 컬럼별 세부설정
