+++ jin/src/main/java/umc/spring/domain/FoodCategory.java
@@ -1,11 +1,8 @@
-package umc.study.domain;
+package umc.spring.domain;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import lombok.*;
-import umc.study.domain.common.BaseEntity;
+import umc.spring.domain.common.BaseEntity;
 
 @Entity
 @Getter
@@ -16,7 +13,8 @@ public class FoodCategory extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Long categoryId;
 
+    @Column(nullable = false, length = 30)
     private String name;
 }
