+++ jin/src/main/java/umc/spring/domain/enums/MemberStatus.java
@@ -1,4 +1,4 @@
-package umc.study.domain.enums;
+package umc.spring.domain.enums;
 
 public enum MemberStatus {
     ACTIVE, INACTIVE
