+++ jin/src/main/java/umc/spring/domain/enums/SocialType.java
@@ -1,4 +1,4 @@
-package umc.study.domain.enums;
+package umc.spring.domain.enums;
 
 public enum SocialType {
     KAKAO, GOOGLE, NAVER, APPLE
