+++ jin/src/main/java/umc/spring/domain/enums/StoreCategory.java
@@ -1,0 +1,5 @@
+package umc.spring.domain.enums;
+
+public enum StoreCategory {
+    KOREAN, CHINESE, JAPANESE, WESTERN, FASTFOOD
+}
