--- jin/src/main/java/umc/study/domain/mapping/Alarm.java
@@ -1,33 +1,0 @@
-package umc.study.domain.mapping;
-
-
-import jakarta.persistence.*;
-import lombok.*;
-import umc.study.domain.User;
-import umc.study.domain.common.BaseEntity;
-
-@Entity
-@Getter
-@Builder
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@AllArgsConstructor
-public class Alarm extends BaseEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    private String message;
-
-    private Boolean isRead;
-
-    // 연관관계 편의 메서드
-    public void setUser(User user) {
-        this.user = user;
-        user.getUserAlarmList().add(this);
-    }
-}
