+++ jin/src/main/resources/application.yml
@@ -1,0 +1,21 @@
+spring:
+  application:
+    name: spring
+  datasource:
+    url: jdbc:mysql://localhost:3306/umc
+    username: root
+    password: 3724
+    driver-class-name: com.mysql.cj.jdbc.Driver
+  sql:
+    init:
+      mode: never
+  jpa:
+    properties:
+      hibernate:
+        dialect: org.hibernate.dialect.MySQL8Dialect
+        show_sql: true
+        format_sql: true
+        use_sql_comments: true
+        hbm2ddl:
+          auto: update
+        default_batch_fetch_size: 1000
